Split MainServer() and MainServerExit into separate files.  Split interface into separate files for windows and linux
