+++ blakserv/main.c
@@ -22,8 +22,6 @@
 
 /* local function prototypes */
 void MainUsage();
-void MainServer();
-void MainExitServer();
 
 DWORD main_thread_id;
 
@@ -44,130 +42,12 @@ int WINAPI WinMain(HINSTANCE hInstance,HINSTANCE hPrev_instance,char *command_li
 
 int main(int argc, char **argv)
 {
+    main_thread_id = pthread_self();
 	MainServer();
 	return 0;
 }
 
 #endif
-
-void MainServer()
-{
-	InitInterfaceLocks(); 
-	
-	InitInterface(); /* starts a thread with the window */
-
-	InitMemory(); /* memory needs channels in general, but need to start before config,
-	so just be careful. */
-	
-	InitConfig();
-	LoadConfig();		/* must be nearly first since channels use it */
-
-	InitDebug();
-	
-	InitChannelBuffer();
-	
-	OpenDefaultChannels();
-
-#ifdef BLAK_PLATFORM_WINDOWS
-	if (ConfigBool(MYSQL_ENABLED))
-	{
-		lprintf("Starting MySQL writer");
-		MySQLInit(ConfigStr(MYSQL_HOST), ConfigStr(MYSQL_USERNAME), ConfigStr(MYSQL_PASSWORD), ConfigStr(MYSQL_DB));
-	}
-#endif
-
-	lprintf("Starting %s\n",BlakServLongVersionString());
-
-	InitClass();
-	InitMessage();
-	InitObject();
-	InitList();
-	InitTimer();
-	InitSession();
-	InitResource();
-	InitRooms();
-	InitString();
-	InitUser();
-	InitAccount();
-	InitNameID();
-	InitDLlist();   
-	InitSysTimer();
-	InitMotd();
-	InitLoadBof();
-	InitTime();
-	InitGameLock();
-	InitBkodInterpret();
-	InitBufferPool();
-	InitTables();
-	AddBuiltInDLlist();
-	LoadMotd();
-	LoadBof();
-	LoadRsc();
-	LoadKodbase();
-	LoadAdminConstants();
-	PauseTimers();
-	
-	if (LoadAll() == True)
-	{
-	/* this loaded_game_msg tells it to disconnect all blakod info about sessions,
-		* that were logged on when we saved */
-		
-		SendTopLevelBlakodMessage(GetSystemObjectID(),LOADED_GAME_MSG,0,NULL);
-		DoneLoadAccounts();
-	}
-	
-	/* these must be after LoadAll and ClearList */
-	InitCommCli(); 
-	InitParseClient(); 
-	InitProfiling();
-	InitAsyncConnections();
-	UpdateSecurityRedbook();
-	UnpauseTimers();
-
-	ServiceTimers(); /* returns if server termiated */
-
-	MainExitServer();
-}
-
-void MainExitServer()
-{
-	lprintf("ExitServer terminating server\n");
-	
-	ExitAsyncConnections();
-	
-	CloseAllSessions(); /* gotta do this before anything, cause it uses kod, accounts */
-	
-	CloseDefaultChannels();
-	
-	ResetLoadMotd();
-	ResetLoadBof();
-	
-	ResetTables();
-	ResetBufferPool();
-	ResetSysTimer();
-	ResetDLlist();
-	ResetNameID();
-	ResetAccount();
-	ResetUser();
-	ResetString();
-	// ExitRooms calls ResetRooms in addition to clearing the array memory.
-	ExitRooms();
-	ResetResource();
-	ResetTimer();
-	ResetList();
-	ResetObject();
-	ResetMessage();
-	ResetClass();
-
-#ifdef BLAK_PLATFORM_WINDOWS
-	if (ConfigBool(MYSQL_ENABLED))
-		MySQLEnd();
-#endif
-	
-	ResetConfig();
-	
-	DeleteAllBlocks();
-}
 
 // This function keeps the necessary calls to reset and reinit game data
 // in one place, to reduce the chance of errors if modifying it. Used by
