+++ blakserv/main_linux.h
@@ -1,0 +1,7 @@
+#ifndef MAIN_LINUX_H
+#define MAIN_LINUX_H
+
+ void MainServer();
+ void MainExitServer();
+
+#endif
