+++ blakserv/main_windows.h
@@ -1,0 +1,7 @@
+#ifndef MAIN_WINDOWS_H
+#define MAIN_WINDOWS_H
+
+ void MainServer();
+ void MainExitServer();
+
+#endif
