+++ blakserv/makefile.linux
@@ -17,7 +17,7 @@ BLAKSERVRUNDIR = $(TOPDIR)/run/server
 
 # -Wno-write-strings: turns off deprecated warning for const char*
 #   strings passed as char* (Should be properly fixed)
-CFLAGS = -I $(BLAKINCLUDEDIR) -x c++ -DBLAK_PLATFORM_LINUX -Wno-write-strings
+CFLAGS = -I $(BLAKINCLUDEDIR) -g -x c++ -DBLAK_PLATFORM_LINUX -Wno-write-strings
 
 
 BLAKSERVLINKFLAGS = -lpthread -lrt
@@ -28,6 +28,7 @@ LIBS =
 
 OBJS =  \
 	$(OUTDIR)/main.obj \
+	$(OUTDIR)/main_linux.obj \
 	$(OUTDIR)/loadkod.obj \
 	$(OUTDIR)/class.obj \
 	$(OUTDIR)/message.obj \
@@ -96,7 +97,7 @@ OBJS =  \
 	$(OUTDIR)/linux-critical_section.obj \
 	$(OUTDIR)/interface_linux.obj \
 	$(OUTDIR)/mutex_linux.obj \
-    $(OUTDIR)/tmsgloop_linux.obj \
+    $(OUTDIR)/thdmsgqueue_linux.obj \
 
 
 all : makedirs $(OUTDIR)/blakserv
