+++ blakserv/thdmsgqueue_windows.c
