+++ blakserv/thdmsgqueue_windows.h
